Update: (SurveysCore) Move the `SurveyError` to Surveys directory in OBAKitCore
