+++ OBAKitCore/Network/Operations/NetworkOperation.swift
@@ -108,20 +108,3 @@ public enum APIError: Error, LocalizedError {
         }
     }
 }
-
-public enum SurveyError: Error, LocalizedError {
-
-    case serviceUnavailable
-
-    case missingUpdatePath
-
-    public var errorDescription: String? {
-        switch self {
-        case .serviceUnavailable:
-            return Strings.surveyServiceNotAvailable
-        case .missingUpdatePath:
-            return Strings.surveyMissingUpdatePath
-        }
-    }
-
-}
